Initial implementation of entity specific token flow enrichment
 * Refactor token-flow implementation into smaller code units
 * Rename existing `createTokenFlowEnricher` function into
`createTokenFlowFullUtteranceEnricher`
 * Add new `createTokenFlowEntityEnricher` which will only enrich
entities from the `RecognizedUtterance` that are selected using a
user supplied  selector function
