+++ packages/tokenflow-integration/src/types.ts
@@ -1,0 +1,19 @@
+import { BasicEntity } from "conversation-processor";
+import { CompositeToken, PID } from "token-flow";
+
+export const ENTITY: unique symbol = Symbol("ENTITY");
+
+export type ENTITY = typeof ENTITY;
+
+export interface TokenFlowEntity extends BasicEntity {
+    type: "token-flow";
+    $raw: any;
+    name: string;
+    pid: PID;
+}
+
+export interface EntityToken extends CompositeToken {
+    type: ENTITY;
+    pid: PID;
+    name: string;
+}
